chore: 自動ツールスイッチャとして独立したクラスを実装
